dns: add parallel resolution test
